Renaming mixnet settings to mixnet contract settings
