+++ contracts/mixnet/src/lib.rs
@@ -4,7 +4,7 @@
 pub mod contract;
 pub mod error;
 pub mod gateways;
-pub mod mixnet_params;
+pub mod mixnet_contract_settings;
 pub mod mixnodes;
 pub mod query_support;
 mod rewards;
